+++ src/lazarus/ide/idehelpmanager.pas
@@ -1737,5 +1737,6 @@ begin
   Result:=ConvertCodePosToPascalHelpContext(@CodePos);
 end;
 
+
 end.